#282 Updated and added tests
